Add missing semicolon in ara-classes-PlaypenSans.fea
Looking into issues with compiling the FEA for this project; it is the first time I have encountered FEA `include` statements inside `feature` blocks.

After getting that working I ran into this: the last statement in this file is missing a semicolon. I'm not sure why fonttools is accepting this, but it might be something silly like that the semicolon after the `include` statement is intact after the concatenation? In any case the spec is clear that there should be a semicolon here.